add user ownership of bottles

+++ app/controllers/api/bottles_controller.rb
@@ -16,9 +16,9 @@ class Api::BottlesController < ApplicationController
   def create
     # :Parameters {"bottle"=>{"name"=>"Macallan 15", "description"=>"Macallan 15", "distillery_id"=>"3", "age"=>"15", "release_year"=>"undefined", "price"=>"100" }, 
     # "photo"=>#<ActionDispatch::Http::UploadedFile:0x00007fc3d422afc0 @tempfile=#<Tempfile:/var/folders/lv/n4scx8rs7lxbnbtdf9v6c4380000gn/T/RackMultipart20190804-43836-1fxp7zt.jpg>, @original_filename="macallan_12_by-TWE.jpg", @content_type="image/jpeg", @headers="Content-Disposition: form-data; name=\"bottle[photo]\"; filename=\"macallan_12_by-TWE.jpg\"\r\nContent-Type: image/jpeg\r\n">}, "format"=>:json, "controller"=>"api/bottles", "action"=>"create"} permitted: false>
-    
+
     if params[:bottle][:age] != ''
-      params[:bottle][:age]  = params[:bottle][:age] .to_i
+      params[:bottle][:age]  = params[:bottle][:age].to_i
     else 
       params[:bottle][:age]  = nil
     end
@@ -32,10 +32,13 @@ class Api::BottlesController < ApplicationController
     if params[:bottle][:price] != '' 
       params[:bottle][:price] = params[:bottle][:price].to_f
     end
-    
+
     @bottle = Bottle.new(bottle_params)
+    debugger
+    @bottle.user_id = current_user.id
     
     if @bottle.save
+      debugger 
       render :show
     else
       render json: @bottle.errors.full_messages, status: 422
@@ -43,8 +46,9 @@ class Api::BottlesController < ApplicationController
   end
 
   def update
-    @bottle = Bottle.find_by(id: params[:id])
-    
+    # @bottle = Bottle.find_by(id: params[:id])
+    debugger
+    @bottle = current_user.bottles.find(params[:id])
   
     if @bottle.update!(bottle_params)
       render :show
@@ -54,17 +58,20 @@ class Api::BottlesController < ApplicationController
   end
 
   def destroy
-    # bottle = current_user.bottles.find_by(id: params[:id])
+    bottle = current_user.bottles.find_by(id: params[:id])
 
-    # if bottle
-    bottle.destroy 
-    render json: ['Bottle successfully removed']
+    if bottle
+      bottle.destroy 
+      render json: ['Bottle successfully removed']
+    else
+      render json: ['You can only delete your own bottles']
+    end
   end
 
   private
 
   def bottle_params
-    params.require(:bottle).permit(:name, :description, :distillery_id, :age, :release_year, :price, :photo)
+    params.require(:bottle).permit(:name, :description, :distillery_id, :age, :release_year, :price, :photo, :user_id)
   end
 
 end
